requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
jinja2==3.1.3
python-dotenv==1.0.1
Pillow==10.2.0
google-generativeai==0.3.2
